+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/PidRegulator.kt
@@ -1,4 +1,4 @@
-package center.sciprog.controls.devices.misc
+package space.kscience.controls.constructor
 
 import kotlinx.coroutines.Job
 import kotlinx.coroutines.delay
@@ -8,28 +8,27 @@ import kotlinx.coroutines.sync.Mutex
 import kotlinx.coroutines.sync.withLock
 import kotlinx.datetime.Clock
 import kotlinx.datetime.Instant
-import space.kscience.controls.api.DeviceLifecycleState
 import space.kscience.controls.spec.DeviceBySpec
 import kotlin.time.Duration
 import kotlin.time.Duration.Companion.milliseconds
 import kotlin.time.DurationUnit
 
 /**
- * PID controller on top of a [Regulator]
+ * A drive with PID regulator
  */
 public class PidRegulator(
-    public val regulator: Regulator,
+    public val drive: Drive,
     public val kp: Double,
     public val ki: Double,
     public val kd: Double,
-    private val dt: Duration = 0.5.milliseconds,
+    private val dt: Duration = 1.milliseconds,
     private val clock: Clock = Clock.System,
-) : DeviceBySpec<Regulator>(Regulator, regulator.context), Regulator {
+) : DeviceBySpec<Regulator>(Regulator, drive.context), Regulator {
 
-    override var target: Double = regulator.target
+    override var target: Double = drive.position
 
     private var lastTime: Instant = clock.now()
-    private var lastRegulatorTarget: Double = target
+    private var lastPosition: Double = target
 
     private var integral: Double = 0.0
 
@@ -39,10 +38,7 @@ public class PidRegulator(
 
 
     override suspend fun onStart() {
-        if(regulator.lifecycleState == DeviceLifecycleState.STOPPED){
-            regulator.start()
-        }
-        regulator.start()
+        drive.start()
         updateJob = launch {
             while (isActive) {
                 delay(dt)
@@ -51,13 +47,13 @@ public class PidRegulator(
                     val delta = target - position
                     val dtSeconds = (realTime - lastTime).toDouble(DurationUnit.SECONDS)
                     integral += delta * dtSeconds
-                    val derivative = (regulator.target - lastRegulatorTarget) / dtSeconds
+                    val derivative = (drive.position - lastPosition) / dtSeconds
 
                     //set last time and value to new values
                     lastTime = realTime
-                    lastRegulatorTarget = regulator.target
+                    lastPosition = drive.position
 
-                    regulator.target = regulator.position + kp * delta + ki * integral + kd * derivative
+                    drive.force = kp * delta + ki * integral + kd * derivative
                 }
             }
         }
@@ -67,129 +63,6 @@ public class PidRegulator(
         updateJob?.cancel()
     }
 
-    override val position: Double get() = regulator.position
+    override val position: Double get() = drive.position
 
 }
-
-
-//
-//interface PidRegulator : Device {
-//    /**
-//     * Proportional coefficient
-//     */
-//    val kp: Double
-//
-//    /**
-//     * Integral coefficient
-//     */
-//    val ki: Double
-//
-//    /**
-//     * Differential coefficient
-//     */
-//    val kd: Double
-//
-//    /**
-//     * The target value for PID
-//     */
-//    var target: Double
-//
-//    /**
-//     * Read current value
-//     */
-//    suspend fun read(): Double
-//
-//    companion object : DeviceSpec<PidRegulator>() {
-//        val target by property(MetaConverter.double, PidRegulator::target)
-//        val value by doubleProperty { read() }
-//    }
-//}
-//
-///**
-// *
-// */
-//class VirtualPid(
-//    context: Context,
-//    override val kp: Double,
-//    override val ki: Double,
-//    override val kd: Double,
-//    val mass: Double,
-//    override var target: Double = 0.0,
-//    private val dt: Duration = 0.5.milliseconds,
-//    private val clock: Clock = Clock.System,
-//) : DeviceBySpec<PidRegulator>(PidRegulator, context), PidRegulator {
-//
-//    private val mutex = Mutex()
-//
-//
-//    private var lastTime: Instant = clock.now()
-//    private var lastValue: Double = target
-//
-//    private var value: Double = target
-//    private var velocity: Double = 0.0
-//    private var acceleration: Double = 0.0
-//    private var integral: Double = 0.0
-//
-//
-//    private var updateJob: Job? = null
-//
-//    override suspend fun onStart() {
-//        updateJob = launch {
-//            while (isActive) {
-//                delay(dt)
-//                mutex.withLock {
-//                    val realTime = clock.now()
-//                    val delta = target - value
-//                    val dtSeconds = (realTime - lastTime).toDouble(DurationUnit.SECONDS)
-//                    integral += delta * dtSeconds
-//                    val derivative = (value - lastValue) / dtSeconds
-//
-//                    //set last time and value to new values
-//                    lastTime = realTime
-//                    lastValue = value
-//
-//                    // compute new value based on velocity and acceleration from the previous step
-//                    value += velocity * dtSeconds + acceleration * dtSeconds.pow(2) / 2
-//
-//                    // compute new velocity based on acceleration on the previous step
-//                    velocity += acceleration * dtSeconds
-//
-//                    //compute force for the next step based on current values
-//                    acceleration = (kp * delta + ki * integral + kd * derivative) / mass
-//
-//
-//                    check(value.isFinite() && velocity.isFinite()) {
-//                        "Value $value is not finite"
-//                    }
-//                }
-//            }
-//        }
-//    }
-//
-//    override fun onStop() {
-//        updateJob?.cancel()
-//        super<PidRegulator>.stop()
-//    }
-//
-//    override suspend fun read(): Double = value
-//
-//    suspend fun readVelocity(): Double = velocity
-//
-//    suspend fun readAcceleration(): Double = acceleration
-//
-//    suspend fun write(newTarget: Double) = mutex.withLock {
-//        require(newTarget.isFinite()) { "Value $newTarget is not valid" }
-//        target = newTarget
-//    }
-//
-//    companion object : Factory<Device> {
-//        override fun build(context: Context, meta: Meta) = VirtualPid(
-//            context,
-//            meta["kp"].double ?: error("Kp is not defined"),
-//            meta["ki"].double ?: error("Ki is not defined"),
-//            meta["kd"].double ?: error("Kd is not defined"),
-//            meta["m"].double ?: error("Mass is not defined"),
-//        )
-//
-//    }
-//}