+++ controls-constructor/src/commonMain/kotlin/space/kscience/controls/constructor/Regulator.kt
@@ -1,17 +1,14 @@
-package center.sciprog.controls.devices.misc
+package space.kscience.controls.constructor
 
-import kotlinx.coroutines.Job
 import space.kscience.controls.api.Device
-import space.kscience.controls.spec.DeviceBySpec
 import space.kscience.controls.spec.DevicePropertySpec
 import space.kscience.controls.spec.DeviceSpec
 import space.kscience.controls.spec.doubleProperty
-import space.kscience.dataforge.context.Context
 import space.kscience.dataforge.meta.transformations.MetaConverter
 
 
 /**
- * A single axis drive
+ * A regulator with target value and current position
  */
 public interface Regulator : Device {
     /**
@@ -29,23 +26,4 @@ public interface Regulator : Device {
 
         public val position: DevicePropertySpec<Regulator, Double> by doubleProperty { position }
     }
-}
-
-/**
- * Virtual [Regulator] with speed limit
- */
-public class VirtualRegulator(
-    context: Context,
-    value: Double,
-    private val speed: Double,
-) : DeviceBySpec<Regulator>(Regulator, context), Regulator {
-
-    private var moveJob: Job? = null
-
-    override var position: Double = value
-        private set
-
-    override var target: Double = value
-
-
 }